// frontend/src/api/models/FoodInstantResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Photo } from './Photo';
export type FoodInstantResponse = {
  photo?: Photo;
  nf_calories?: number;
  food_name?: string;
  tag_id?: string;
  nix_item_id?: string;
};